Move adjustNodeInsertPosition to core package
